% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_name_update}
\alias{check_name_update}
\title{Check taxonomic name updates}
\usage{
check_name_update(resultado, threatenedperu)
}
\arguments{
\item{resultado}{A tibble or data frame returned by
\code{matching_threatenedperu()}, containing at least the column
\code{Matched.Name}.}

\item{threatenedperu}{A tibble or data frame with taxonomic reference
information, containing at least the columns \code{scientific_name}
and \code{taxonomic_status}.}
}
\value{
A character string reporting the number and percentage of
  names that have been taxonomically updated.
}
\description{
This function compares a set of matched names (from the function
\code{matching_threatenedperu()}) against the reference dataset
\code{threatenedperu}. It determines how many of the queried names
are no longer accepted and have been taxonomically updated.
}
\examples{
\dontrun{
resultado <- matching_threatenedperu(c(
  "Haageocereus acranthus subsp. olowinskianus",
  "Tecoma estans var. velutyno",
  "Haageocereus pseudomelanostele subsp. setosus",
  "Cleistocactus pachycladus"
)) |>
  tibble::as_tibble()

check_name_update(resultado, threatenedperu)
}

}
\keyword{internal}
