% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match_infra_rank_within_species_helper}
\alias{direct_match_infra_rank_within_species_helper}
\title{Helper: Direct Match Infraspecific Rank within Species}
\usage{
direct_match_infra_rank_within_species_helper(df, target_df)
}
\arguments{
\item{df}{A tibble containing data for a single matched species.}

\item{target_df}{A tibble representing the threatened species database.}
}
\value{
A tibble with match results.
}
\description{
Helper function that performs the actual matching of infraspecific ranks
for a single matched species.
}
\keyword{internal}
