% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-pipeline.R
\name{.initialize_matching_columns}
\alias{.initialize_matching_columns}
\title{Initialize Matching Columns}
\usage{
.initialize_matching_columns(df, use_infraspecies_2)
}
\arguments{
\item{df}{Dataframe with classified species}

\item{use_infraspecies_2}{Logical, whether Rank 4 is supported}
}
\value{
Dataframe with initialized matching columns
}
\description{
Adds all necessary matching columns to the dataframe with NA defaults.
}
\keyword{internal}
