% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-pipeline.R
\name{.join_threat_info_original}
\alias{.join_threat_info_original}
\title{Join Threat Info for Original Database (Supports Rank 4)}
\usage{
.join_threat_info_original(res_complete, target_df)
}
\description{
Join Threat Info for Original Database (Supports Rank 4)
}
\keyword{internal}
