% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-pipeline.R
\name{.join_threat_info_updated}
\alias{.join_threat_info_updated}
\title{Join Threat Info for Updated Database (Rank 3 max)}
\usage{
.join_threat_info_updated(res_complete, target_df)
}
\description{
Join Threat Info for Updated Database (Rank 3 max)
}
\keyword{internal}
