% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-pipeline.R
\name{.recalculate_rank_infrasp2}
\alias{.recalculate_rank_infrasp2}
\title{Recalculate Rank After Infraspecies Level 2 Match}
\usage{
.recalculate_rank_infrasp2(df)
}
\description{
Recalculate Rank After Infraspecies Level 2 Match
}
\keyword{internal}
