% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match_infraspecies2_within_infraspecies}
\alias{fuzzy_match_infraspecies2_within_infraspecies}
\title{Fuzzy Match Infraspecies Level 2 within Infraspecies Level 1}
\usage{
fuzzy_match_infraspecies2_within_infraspecies(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the species data to be matched.}

\item{target_df}{A tibble representing the threatened species database
containing the reference list of threatened species.}
}
\value{
A tibble with an additional logical column fuzzy_match_infraspecies2,
indicating whether the second infraspecific level was successfully fuzzy
matched within the matched first infraspecific level (`TRUE`) or not (`FALSE`).
}
\description{
This function performs a fuzzy match of the second infraspecific level within
an already matched first infraspecific level from the list of threatened
species in the database. Follows the same pattern as
fuzzy_match_infraspecies_within_species.
}
\keyword{internal}
