% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match_infraspecies_within_species}
\alias{fuzzy_match_infraspecies_within_species}
\title{Fuzzy Match Infraspecies within Species}
\usage{
fuzzy_match_infraspecies_within_species(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the species data to be matched.}

\item{target_df}{A tibble representing the threatened species database containing the reference list of threatened species.}
}
\value{
A tibble with an additional logical column fuzzy_match_infraspecies_within_species, indicating whether the specific infraspecies was successfully fuzzy matched within the matched species (`TRUE`) or not (`FALSE`).
}
\description{
This function performs a fuzzy match of specific infraspecies within an already matched epithet from the list of threatened species in the database.
}
\keyword{internal}
