% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match_infraspecies_within_species_helper}
\alias{fuzzy_match_infraspecies_within_species_helper}
\title{Helper: Fuzzy Match Infraspecific Epithet within Species}
\usage{
fuzzy_match_infraspecies_within_species_helper(df, target_df)
}
\arguments{
\item{df}{A tibble containing data for a single matched species.}

\item{target_df}{A tibble representing the threatened species database.}
}
\value{
A tibble with fuzzy match results and logical indicator.
}
\description{
Helper function that performs fuzzy matching of infraspecific epithets
for a single matched species. The infraspecific rank category must already
be validated.
}
\keyword{internal}
