% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_match.R
\name{get_threat_summary}
\alias{get_threat_summary}
\title{Get threat categories summary}
\usage{
get_threat_summary(splist, target_df = NULL)
}
\arguments{
\item{splist}{A character vector containing species names}

\item{target_df}{A tibble representing the threatened species database}
}
\value{
A tibble with threat category counts
}
\description{
Get a summary of threat categories for a list of species
}
