% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_match.R
\name{is_single_species_threatened}
\alias{is_single_species_threatened}
\title{Check if a single species is threatened}
\usage{
is_single_species_threatened(species_name, target_df = NULL)
}
\arguments{
\item{species_name}{A character string with a single species name}

\item{target_df}{A tibble representing the threatened species database}
}
\value{
A character string with the threat status
}
\description{
Convenience function to check the threat status of a single species
}
